Move local dependency to localpkg folder
